backend/requirements.txt
# ==================== Python Version ====================
# Python 3.9+
# Tested on: Python 3.9.25
# Minimum: Python 3.9.0
# Maximum: Python 3.11.x (Python 3.12+ 미지원 패키지 있음)

# ==================== FastAPI & Web Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0
pydantic==2.4.2
pydantic-core==2.10.1

# ==================== Database ====================
pymysql==1.1.0
cryptography==41.0.7

# ==================== Data Processing ====================
pandas==2.1.3
numpy==1.26.2
openpyxl==3.1.2

# ==================== PDF & Document Processing ====================
reportlab==4.0.7
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.1.0

# ==================== AI & LLM ====================
openai==1.3.7
anthropic==0.7.1
groq==0.4.1
google-generativeai==0.3.1

# ==================== RAG & Vector Store ====================
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.3.1
huggingface-hub==0.20.3
faiss-cpu==1.7.4
transformers==4.35.2
torch==2.1.1

# ==================== HTTP & Networking ====================
requests==2.31.0
httpx==0.25.2
urllib3==2.1.0

# ==================== Utilities ====================
python-dotenv==1.0.0
aiofiles==23.2.1

# ==================== Optional (Development) ====================
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.11.0
# flake8==6.1.0
